feat: Number 생성 및 문자열을 정수처리 로직 구현

+++ src/test/java/calculator/NumberTest.java
@@ -8,7 +8,7 @@ public class NumberTest {
 
     @Test
     void Number_생성() {
-        int num = new Number("1");
+        int num = new Number("1").intValue();
         assertThat(num).isEqualTo(1);
     }
 }
